requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6